# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: SensiLog API
  version: 1.0.0
  description: VALORANT感度・設定管理API
  contact:
    name: SensiLog Team
    url: https://sensilog.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sensilog.com
    description: Production
  - url: http://localhost:3001
    description: Development
...
paths:
...
  /auth/riot/callback:
    post:
      summary: Riot OAuth認証コールバック
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: 認証コード
                state:
                  type: string
                  description: CSRF保護用state
              required:
                - code
      responses:
        200:
          description: 認証成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWTアクセストークン
                  expiresAt:
                    type: string
                    format: date-time
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: トークンリフレッシュ
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: 新しいトークン
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
        401:
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      summary: ユーザー情報取得
      tags: [Authentication]
      security:
        - bearerAuth: []
      responses:
        200:
          description: ユーザー情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        riotPuuid:
          type: string
          description: Riot PUUID
        gameName:
          type: string
          description: Riot ID
        tagLine:
          type: string
          description: Riot Tag
        isAdmin:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - createdAt
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
        code:
          type: string
          description: エラーコード
        details:
          type: object
          description: 詳細情報
      required:
        - error
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 認証が必要
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT